// frontend/src/app/api/models/coleccion-detalle.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ElementoColeccion } from '../models/elemento-coleccion';
import { ImagenPublica } from '../models/imagen-publica';
import { MetadatosEditoriales } from '../models/metadatos-editoriales';
import { Slug } from '../models/slug';
import { TextoEnriquecido } from '../models/texto-enriquecido';
export interface ColeccionDetalle {
  descripcion: TextoEnriquecido;
  elementos: Array<ElementoColeccion>;
  metadatos: MetadatosEditoriales;
  portada: ImagenPublica;
  resumen: string;
  slug: Slug;
  titulo: string;
}