// frontend/src/app/api/models/coleccion-tarjeta.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ImagenPublicaNula } from '../models/imagen-publica-nula';
import { Slug } from '../models/slug';
export interface ColeccionTarjeta {
  numero_elementos: number;
  portada: ImagenPublicaNula;
  resumen: string;
  slug: Slug;
  titulo: string;
}