// frontend/src/app/api/models/itinerario-tarjeta.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Dificultad } from '../models/dificultad';
import { ImagenPublicaNula } from '../models/imagen-publica-nula';
import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
export interface ItinerarioTarjeta {
  destino: RefTaxonomia;
  dificultad: Dificultad;
  duracion_dias: number;
  portada: ImagenPublicaNula;
  slug: Slug;
  titulo: string;
}